Move home-related files to view/com/home
